requirements-fixed.txt
# Core Dependencies - Pinned stable versions
streamlit==1.40.0
python-dotenv==1.0.0

# LangGraph and LangChain - Compatible versions
langgraph==0.2.45
langchain-core==0.3.15
langchain-groq==0.2.1

# LLM Providers
groq==0.11.0
google-generativeai==0.7.2

# Content Extraction (Jina is used via API, no package needed!)
requests==2.32.3
beautifulsoup4==4.12.3

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2

# Audio Recording (Phase 3 STT features)
streamlit-webrtc==0.47.9
pydub==0.25.1

# HTTP/Async Support (required by langchain/groq)
httpx==0.27.2
httpcore==1.0.7

# Pydantic (critical - must be v2 for langchain)
pydantic==2.9.2
pydantic-core==2.23.4
